Users` money are now double-precision.

+++ FlightsReservation.BLL/Entities/DataTransferObjects/UserDtos/UserUpdateDto.cs
@@ -3,5 +3,5 @@
 public class UserUpdateDto : BaseUserDto
 {
     public Guid Id { get; set; }
-    public int Money { get; set; }
+    public double Money { get; set; }
 }
